% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{bagging}
\alias{bagging}
\title{Bagging}
\usage{
bagging(X, y, model, criteria)
}
\arguments{
\item{X}{Whole predictors matrix}

\item{y}{Whole response variables}

\item{model}{The name of model that user prefer}

\item{criteria}{The criteria of the response variable(Bianry or Continuous)}
}
\value{
Predicted values and test statistics of model
}
\description{
This function will promote users to determine the rounds of bootstrap
sampling and perform the specific regression model for those rounds.
Finally, this function will aggregate the final result by averaging
from all the predicted values from previous rounds.
}
\examples{
# Bagging for a specific model
bagging(X,y,"lasso","Continuous")
}
