% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble.learn}
\alias{ensemble.learn}
\title{Ensemble learning}
\usage{
ensemble.learn(X, y, criteria, comb)
}
\arguments{
\item{X}{Whole predictors matrix}

\item{y}{Whole response variables}

\item{criteria}{The criteria of the response variable(Bianry or Continuous)}

\item{comb}{List of names for the model that user prefer}
}
\value{
Predicted values and test statistics of model
}
\description{
This package enables ensemble learning for the data. The function ensemble.
learn takes in X and y, and criteria and comb as input, criteria can be
binary and continuous, comb are the different kinds of regression models that
user wants to perform. For each regression model, it will perform bagging
learning on them, the results from different will be combined into a matrix,
each column is the result of a different regression model. If the criteria
is binary, then the majority voting will be performed by calling the vote.max
function, and result will be returned with values predicted and the accuracy.
The vote.max function chooses the mode of the binary vector. If the criteria
is continuous, then the the mean of matrix of each column will be found and
returned as the predicted values, mean square error(MSR) will be calculated
and it will return the user the MSR and the predicted value.
}
\examples{
#Define list of combination of model
list_samp <- c("lasso","randomforest")
#Ensemble lean
ensemble.learn(X,y,"Continuous",list_samp)
}
