% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_logistic.R
\name{linear.logistic}
\alias{linear.logistic}
\title{Linear & Logistic Regression}
\usage{
linear.logistic(X.train, X.test, y.train, y.test, criteria)
}
\arguments{
\item{X.train}{The train data to build the model}

\item{X.test}{The test data to predict outcomes}

\item{y.train}{The response variable of train data}

\item{y.test}{The true response variable of test data}

\item{criteria}{The criteria of the response variable(Bianry or Continuous)}
}
\value{
Predicted values and test statistics of model
}
\description{
As user input a Matrix of X with predictors and y as response variable,
this function will train model in the way of Linear regression
or logistic regression depending on the criteria of the result.
}
\examples{
#Build linear regression model
linear.logistic(X.train, X.test, y.train, y.test, "Continuous")
#Build logistic regression model
linear.logistic(X.train, X.test, y.train, y.test, "Binary")
}
