% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic.R
\name{my.elastic}
\alias{my.elastic}
\title{Elastic net Regression}
\usage{
my.elastic(X.train, X.test, y.train, y.test, criteria)
}
\arguments{
\item{X.train}{The train data to build the model}

\item{X.test}{The test data to predict outcomes}

\item{y.train}{The response variable of train data}

\item{y.test}{The true response variable of test data}

\item{criteria}{The criteria of the response variable(Bianry or Continuous)}
}
\value{
Predicted values and test statistics of model
}
\description{
This function will enable users to perform elastic regression. Loop through
alpha from 0 to 1 and do cross-validation tasks on each alpha value.
For each alpha value, do elastic regression based on the best lambda
value and compare with each other. Then choose the best elastic model
based on the alpha value.
}
\examples{
#Elastic net regression model
my.elastic(X.train, X.test, y.train, y.test, "Continuous")
}
