% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpred.R
\name{my.k.pred}
\alias{my.k.pred}
\title{K predictors selection}
\usage{
my.k.pred(X, Y, k)
}
\arguments{
\item{X}{Matrix of predictor variables}

\item{Y}{Response variable that can be either continuous or binary}

\item{k}{Number of top predictors to select}
}
\value{
Names of the top k predictors
}
\description{
Select top predictors based on correlation with response variable
}
\examples{
my.k.pred(mtcars[, -1], mtcars[, 1], k = 3)
}
