% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{my.lasso}
\alias{my.lasso}
\title{Lasso Regression}
\usage{
my.lasso(X.train, X.test, y.train, y.test, criteria)
}
\arguments{
\item{X.train}{The train data to build the model}

\item{X.test}{The test data to predict outcomes}

\item{y.train}{The response variable of train data}

\item{y.test}{The true response variable of test data}

\item{criteria}{The criteria of the response variable(Bianry or Continuous)}
}
\value{
Predicted values and test statistics of model
}
\description{
This function will enable users to perform lasso regression. First call
cv.glmnet, do cross-validation and select the best lambda.
Then train the lasso model based on the best lambda.
}
\examples{
#Lasso regression model
my.lasso(X.train, X.test, y.train, y.test, "Continuous")
}
