% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomforest.R
\name{my.randomForest}
\alias{my.randomForest}
\title{Random Forest Model}
\usage{
my.randomForest(X.train, X.test, y.train, y.test, criteria)
}
\arguments{
\item{X.train}{The train data to build the model}

\item{X.test}{The test data to predict outcomes}

\item{y.train}{The response variable of train data}

\item{y.test}{The true response variable of test data}

\item{criteria}{The criteria of the response variable(Bianry or Continuous)}
}
\value{
Predicted values and test statistics of model
}
\description{
As user input a Matrix of X with predictors and y as response variable,
this function will train model in the way of random forest.
}
\examples{
#Build RandomForest model
my.randomForest(X.train, X.test, y.train, y.test, "Continuous")
}
