% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{my.ridge}
\alias{my.ridge}
\title{Ridge Regression}
\usage{
my.ridge(X.train, X.test, y.train, y.test, criteria)
}
\arguments{
\item{X.train}{The train data to build the model}

\item{X.test}{The test data to predict outcomes}

\item{y.train}{The response variable of train data}

\item{y.test}{The true response variable of test data}

\item{criteria}{The criteria of the response variable(Bianry or Continuous)}
}
\value{
Predicted values and test statistics of model
}
\description{
This function will enable users to perform ridge regression. First call
cv.glmnet, do cross-validation and select the best lambda.
Then train the ridge model based on the best lambda.
}
\examples{
#Build ridge model
my.ridge(X.train, X.test, y.train, y.test, "Continuous")
}
