% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{simpleEnsembleGroup3}
\alias{simpleEnsembleGroup3}
\title{Main function of the project}
\usage{
simpleEnsembleGroup3(X, y, model)
}
\arguments{
\item{X}{Whole predictors matrix}

\item{y}{Whole response variables}

\item{model}{The name of model that user prefer}
}
\value{
Predicted values and test statistics of model
}
\description{
The main function is mainly served as user interface, it allows users to
input their data and their command of what kinds of regression functions to
implement on the data, moreover, it also checks the integrity of the data,
it checks if the predictor data (mainly X) are in dataframe form, if not it
will prompt the user to input it into one; it checks if the data has missing
value; It checks if the regression model that user wanted to implement
matches the one our package support ("linear&logistic", "ridge", "lasso",
"elastic", "randomforest"), if it matches, it will call the “function”
package and perform corresponding operation; It checks if the predictors and
outcome has the same size; It will prompt user to change the outcome data
into 0 or 1 if data type is binary.  It will prompt the user to choose the
most informative predictors if the predictors have greater number than sample
size; It will convert categorical data to dummy variables; It will also allow
users to choose if they want to perform bagging learning on the data, if yes,
the bagging function in the bagging package will be called. Lastly, it also
prompts users to choose if they want to perform ensemble learning on the
data, if yes, the ensemble.learn function in the ensemble package will be
called. The input for the yes and no questions from users are managed by the
interactive.check function in the “function” package.
}
\examples{
sipmleEnsembleGroup3(X,y,"ridge")
}
